:truck: Move storybook samples
