+++ client/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "Awesome Project Template",
+  "name": "Awesome Project Template Sample",
   "icons": [
     {
       "src": "favicon.ico",
