+++ client/src/components/Button/Button.js
@@ -5,7 +5,7 @@ import "./button.css";
 /**
  * Primary UI component for user interaction
  */
-export const Button = ({ primary, backgroundColor, size, label, ...props }) => {
+const Button = ({ primary, backgroundColor, size, label, ...props }) => {
   const mode = primary
     ? "storybook-button--primary"
     : "storybook-button--secondary";
@@ -52,3 +52,5 @@ Button.defaultProps = {
   size: "medium",
   onClick: undefined,
 };
+
+export default Button;
