+++ client/src/components/Button/Button.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Button } from "./Button";
+import Button from "./Button";
 
 export default {
   title: "Example/Button",
