+++ client/src/components/Header/Header.js
@@ -1,10 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { Button } from "./Button";
+import Button from "../Button";
 import "./header.css";
 
-export const Header = ({ user, onLogin, onLogout, onCreateAccount }) => (
+const Header = ({ user, onLogin, onLogout, onCreateAccount }) => (
   <header>
     <div className="wrapper">
       <div>
@@ -60,3 +60,5 @@ Header.propTypes = {
 Header.defaultProps = {
   user: null,
 };
+
+export default Header;
