+++ client/src/components/Header/Header.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Header } from "./Header";
+import Header from "./Header";
 
 export default {
   title: "Example/Header",
