+++ client/src/pages/Page/Page.js
@@ -1,10 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { Header } from "./Header";
+import Header from "../../components/Header";
 import "./page.css";
 
-export const Page = ({ user, onLogin, onLogout, onCreateAccount }) => (
+const Page = ({ user, onLogin, onLogout, onCreateAccount }) => (
   <article>
     <Header
       user={user}
@@ -91,3 +91,5 @@ Page.propTypes = {
 Page.defaultProps = {
   user: null,
 };
+
+export default Page;
