+++ client/src/pages/Page/Page.stories.js
@@ -1,7 +1,7 @@
 import React from "react";
 
-import { Page } from "./Page";
-import * as HeaderStories from "./Header.stories";
+import Page from "./Page";
+import * as HeaderStories from "../../components/Header/Header.stories";
 
 export default {
   title: "Example/Page",
