+++ package.json
@@ -34,6 +34,7 @@
     "start": "node server.js",
     "server": "nodemon server.js",
     "client": "cd client && npm start",
+    "storybook": "cd client && npm run storybook",
     "dev": "concurrently \"npm run server\" \"npm run client\"",
     "build": "cd client && npm run build && npm run build-storybook",
     "postinstall": "cd client && npm install"
